server/requirements.txt
asgiref==3.4.1
Django==3.2.9
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
Markdown==3.3.4
mqtt==0.0.1
numpy==1.19.5
paho-mqtt==1.6.1
psycopg2==2.9.2
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
